Added bitboard board representation and masks
For this project, I will be using the bitboard board
layout method. This involves 12, 64-bit unsigned ints
each corresponding to the 12 types of pieces on a chessboard
(color-sensitive). the LSB of this int corresponds to a1 and
MSB to h8. 1 = presence of piece in this square, 0 = piece
not in this square.